Several small bugfixes/updates

+++ Httbar/scripts/produce_morph_files_tanb_mA.py
@@ -22,7 +22,8 @@ parser.add_argument('jobid')
 parser.add_argument('mA', type=int)
 parser.add_argument('tanb')
 parser.add_argument('input_sushi')
-parser.add_argument('--blind', action='store_true')
+parser.add_argument('--noblind', action='store_true')
+parser.add_argument('--norm', action='store_true')
 # parser.add_argument('inputs_bkg', help='comma separated list of bkg input files. Same order required as for signal files')
 # parser.add_argument('outfile')
 args = parser.parse_args()
@@ -85,20 +86,21 @@ with open(args.input_sushi) as sushi_pkl:
 	print '\n\nRunning LIMIT\n\n'
 	syscall((
 			'combineTool.py -M Asymptotic -d A_{}_{}_H_{}_{}/workspace.root --there -n'
-			' .limit --minimizerTolerance=0.0001 --minimizerStrategy=2').format(
-			val2name(widthA), mA, val2name(widthH), mH, '--run blind' if args.blind else ''
+			' .limit --minimizerTolerance=0.0001 --minimizerStrategy=2 {}').format(
+			val2name(widthA), mA, val2name(widthH), mH, '' if args.noblind else '--run blind -t -1'
 			))
 	syscall((
 			'combineTool.py -M CollectLimits '
 			'A_{}_{}_H_{}_{}/higgsCombine.limit.Asymptotic.mH120.root').format(
-			val2name(widthA), mA, val2name(widthH), mH, '--run blind' if args.blind else ''
+			val2name(widthA), mA, val2name(widthH), mH
 			))
 shutil.move('limits.json', 'mA%d_tanb%s.json' % (mA, tanb))
-shutil.rmtree(
-        'A_{}_{}_H_{}_{}'.format(
-                val2name(widthA), mA, val2name(widthH), mH, '--run blind' if args.blind else ''
-                )
-        )
-for fname in glob('*.root'):
-	os.remove(fname)
+if not args.norm:
+	shutil.rmtree(
+		'A_{}_{}_H_{}_{}'.format(
+			val2name(widthA), mA, val2name(widthH), mH
+			)
+		)
+	for fname in glob('*.root'):
+		os.remove(fname)
